requirements.txt
streamlit==1.51.0
reportlab==4.4.4
pandas==2.3.3
supabase>=2.0.0
python-dotenv>=1.0.0
pdfplumber==0.11.4

